requirements.txt
absl-py==0.9.0
aniso8601==8.0.0
astor==0.8.1
attrs==19.3.0
cachetools==4.0.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cryptography==2.8
firebase==3.0.1
Flask==1.1.1
flask-restplus==0.13.0
gast==0.2.2
gcloud==0.18.3
google-api-core==1.16.0
google-auth==1.11.3
google-auth-oauthlib==0.4.1
google-cloud-core==1.3.0
google-cloud-storage==1.26.0
google-cloud-vision==1.0.0
google-pasta==0.2.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
grpcio==1.27.2
gunicorn==20.0.4
h5py==2.10.0
httplib2==0.17.0
idna==2.9
importlib-metadata==1.5.0
itsdangerous==1.1.0
Jinja2==2.11.1
jsonschema==3.2.0
jwcrypto==0.7
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.2.1
MarkupSafe==1.1.1
numpy==1.18.2
oauth2client==4.1.3
oauthlib==3.1.0
opt-einsum==3.2.0
pandas==1.0.3
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycrypto==2.6.1
pyrsistent==0.15.7
python-dateutil==2.8.1
python-jwt==3.2.6
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.0
scipy==1.4.1
six==1.14.0
sseclient==0.0.26
tensorboard==2.1.1
tensorflow==2.1.0
tensorflow-estimator==2.1.0
tensorflow-hub==0.7.0
termcolor==1.1.0
urllib3==1.25.8
Werkzeug==1.0.0
wrapt==1.12.1
zipp==3.1.0
